fix: 金額はfloat型だった

+++ server/db/migrate/20230621154840_create_debt_records.rb
@@ -4,7 +4,7 @@ class CreateDebtRecords < ActiveRecord::Migration[7.0]
       t.references :lending_user, null: false, foreign_key: {to_table: :users}, index: false
       t.references :borrowing_user, null: false, foreign_key: {to_table: :users}
       t.references :expense, null: false, foreign_key: true
-      t.integer :amount_of_money, null: false
+      t.float :amount_of_money, null: false
 
       t.timestamps
